+++ packages/nextjs/test/integration/tsconfig.test.json
@@ -1,5 +1,5 @@
 {
-  "extends": "./tsconfig.json",
+  "extends": "../../tsconfig.test.json",
 
   "include": ["test/**/*"],
